patch lsbuf lef + more cpf fixes
Patch LEF for level shifters to change class to block since this is not
possible at runtime. Fix selection of cells in CPF, and reload after IO
cell insertion to make sure physical-only cells are applied.

We might want to drop the IO power domain so all the IO cells/fillers go
on default, and not need to deal with reloading the CPF...